### 2026-03-16 #10201 @editorialbot
[PRE REVIEW]: Geomorphometry.jl: Analyzing and visualizing the shape of the Earth in Julia
**Submitting author:** @evetion (<a href="http://orcid.org/0000-0001-8758-3939">Maarten Pronk</a>)
**Repository:** https://github.com/Deltares/Geomorphometry.jl
**Branch with paper.md** (empty if default branch): feat/paper
**Version:** v0.7.3
