### 2026-03-27 #10201 @elbeejay
@r-barnes, @boriskaus, @vlandau, @mfsch - would any of you be willing to review the submission *Geomorphometry.jl: Analyzing and visualizing the shape of the Earth in Julia* for JOSS?

You've all interacted with JOSS as reviewers and/or authors in the past, and have experience with either geoscience/earth science, Julia, or both, which is why I'm reaching out to you as candidate reviewers. Right now we're asking for reviewers to complete their initial reviews within 6 weeks of the review issue beginning. Please contact me if you have any procedural questions about JOSS, otherwise please take a look at this submission and consider signing up to review.

Thanks!
Jay
