+++ dev/assets/reference.md.D66P6WN2.js
@@ -1,13 +1,0 @@
-import{_ as r,c as n,a5 as o,j as s,a,G as t,w as l,B as p,o as h}from"./chunks/framework.De9caxpi.js";const xe=JSON.parse('{"title":"","description":"","frontmatter":{},"headers":[],"relativePath":"reference.md","filePath":"reference.md","lastUpdated":null}'),d={name:"reference.md"},k={class:"jldocstring custom-block",open:""},m={class:"jldocstring custom-block",open:""},c={class:"jldocstring custom-block",open:""},y={class:"jldocstring custom-block",open:""},g={class:"jldocstring custom-block",open:""},u={class:"jldocstring custom-block",open:""},b={class:"jldocstring custom-block",open:""},E={class:"jldocstring custom-block",open:""},f={class:"jldocstring custom-block",open:""},F={class:"jldocstring custom-block",open:""},j={class:"jldocstring custom-block",open:""},C={class:"jldocstring custom-block",open:""},v={class:"jldocstring custom-block",open:""},x={class:"jldocstring custom-block",open:""},G={class:"jldocstring custom-block",open:""},A={class:"jldocstring custom-block",open:""},D={class:"jldocstring custom-block",open:""},T={class:"jldocstring custom-block",open:""},M={class:"jldocstring custom-block",open:""},w={class:"jldocstring custom-block",open:""},B={class:"jldocstring custom-block",open:""},L={class:"jldocstring custom-block",open:""},R={class:"jldocstring custom-block",open:""},z={class:"jldocstring custom-block",open:""},I={class:"jldocstring custom-block",open:""},O={class:"jldocstring custom-block",open:""},S={class:"jldocstring custom-block",open:""},P={class:"jldocstring custom-block",open:""},N={class:"jldocstring custom-block",open:""},U={class:"jldocstring custom-block",open:""},V={class:"jldocstring custom-block",open:""},H={class:"jldocstring custom-block",open:""},q={class:"jldocstring custom-block",open:""},W={class:"jldocstring custom-block",open:""},Z={class:"jldocstring custom-block",open:""},$={class:"jldocstring custom-block",open:""},J={class:"jldocstring custom-block",open:""},K={class:"jldocstring custom-block",open:""},Q={class:"jldocstring custom-block",open:""},X={class:"jldocstring custom-block",open:""},Y={class:"jldocstring custom-block",open:""},_={class:"jldocstring custom-block",open:""},ee={class:"jldocstring custom-block",open:""},se={class:"jldocstring custom-block",open:""},ie={class:"jldocstring custom-block",open:""},te={class:"jldocstring custom-block",open:""},ae={class:"jldocstring custom-block",open:""},le={class:"jldocstring custom-block",open:""},oe={class:"jldocstring custom-block",open:""},re={class:"jldocstring custom-block",open:""},ne={class:"jldocstring custom-block",open:""},pe={class:"jldocstring custom-block",open:""},he={class:"jldocstring custom-block",open:""},de={class:"jldocstring custom-block",open:""},ke={class:"jldocstring custom-block",open:""},me={class:"jldocstring custom-block",open:""},ce={class:"jldocstring custom-block",open:""},ye={class:"jldocstring custom-block",open
[…]
