### 2026-04-05 #10201 @elbeejay
@martibosch, @AlexanderJuestel -  would either of you be willing to review the submission *Geomorphometry.jl: Analyzing and visualizing the shape of the Earth in Julia* for JOSS?

You've published in JOSS in the past, and have experience with geoscience/earth science, which is why I'm reaching out to you as candidate reviewers. Right now we're asking for reviewers to complete their initial reviews within 6 weeks of the review issue beginning. Please contact me if you have any procedural questions about JOSS, otherwise please take a look at this submission and consider signing up to review.

Thanks!
Jay
