### 2026-04-10 #10395 @editorialbot
[REVIEW]: Geomorphometry.jl: Analyzing and visualizing the shape of the Earth in Julia
**Submitting author:** @evetion (<a href="http://orcid.org/0000-0001-8758-3939">Maarten Pronk</a>)
**Repository:** https://github.com/Deltares/Geomorphometry.jl
**Branch with paper.md** (empty if default branch): feat/paper
**Version:** v0.7.3

@boriskaus, @shipengcheng1230, @Alexander-Barth, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
