### 2026-04-14 #10201 @elbeejay
@loeffko, @alextbradley, @anowacki, @shipengcheng1230, @Oddan - would you be willing to review the submission *Geomorphometry.jl: Analyzing and visualizing the shape of the Earth in Julia* for JOSS?

You've all published Julia software packages in JOSS, and work in the broadly defined field of the earth sciences, which is why I'm reaching out to you as candidate reviewers. Please contact me if you have any procedural questions about JOSS, otherwise please take a look at this submission and consider signing up to review.

Thanks!
Jay
