### 2026-04-14 #10395 @elbeejay
@shipengcheng1230 and @Alexander-Barth, thank you both again for agreeing to review this submission for JOSS. This is the official review issue, please read over the first comment in this issue thread for links and instructions on how to generate your review checklist and conduct your review.

Please do not hesitate to reach out to me with any questions about the process.

Thanks again,
Jay
