### 2026-04-15 Deltares/Geomorphometry.jl#69 @Alexander-Barth
Unavailable links
Thanks @evetion for sharing this nice package :-)

The "Get Started" button (not the "Getting Started" menu) on this page (and the corresponding dev page) is currently unavailable:

https://deltares.github.io/Geomorphometry.jl/stable/

This leads to this broken link:

https://deltares.github.io/Geomorphometry.jl/stable/tutorials/installation (404 error)

it should probably link to this instead:

https://deltares.github.io/Geomorphometry.jl/stable/installation

The codecov badge in the README (https://github.com/Deltares/Geomorphometry.jl) reports unknown coverage. Having the code coverage information would be quite useful.

Reference: https://github.com/openjournals/joss-reviews/issues/10395
