### 2026-04-17 9a7d60b
Update installation link in index.md
