### 2026-04-17 Deltares/Geomorphometry.jl#69 @evetion
Thanks for reviewing! This should be fixed in https://github.com/Deltares/Geomorphometry.jl/commit/9a7d60b7081288c28645717b53c736cfd2467659, although will only show up on dev for now (I will manually fix it in the gh-pages branch for the previous release(s). Codecov is up as well.
