+++ previews/PR70/assets/todo.md.BcTZTMst.lean.js
@@ -1,0 +1,1 @@
+import{_ as t,c as l,j as e,a as n,o as r}from"./chunks/framework.CT0Wd1gt.js";const m=JSON.parse('{"title":"Planned features","description":"","frontmatter":{},"headers":[],"relativePath":"todo.md","filePath":"todo.md","lastUpdated":null}'),o={name:"todo.md"};function s(d,a,u,i,f,p){return r(),l("div",null,a[0]||(a[0]=[e("h1",{id:"Planned-features",tabindex:"-1"},[n("Planned features "),e("a",{class:"header-anchor",href:"#Planned-features","aria-label":'Permalink to "Planned features {#Planned-features}"'},"​")],-1),e("p",null,"The following features are planned for future releases:",-1),e("ul",null,[e("li",null,[e("p",null,"Classification metrics such as geomorphons")]),e("li",null,[e("p",null,"More support for multiresolution")]),e("li",null,[e("p",null,"Integration with DEM datasets/tiles")])],-1)]))}const h=t(o,[["render",s]]);export{m as __pageData,h as default};

[…]
