### 2026-04-23 Deltares/Geomorphometry.jl#69 @Alexander-Barth
Thanks a lot! I will follow the dev pages from now on.
https://deltares.github.io/Geomorphometry.jl/dev/
(no need to fix the older pages for me).
