### 2026-05-18 b32b16e
Make examples reproducable.
