### 2026-05-18 3a21721
Make examples reproducable.
