+++ previews/PR74/assets/todo.md.BEAJwQu0.lean.js
@@ -1,0 +1,1 @@
+import{_ as t,c as l,j as e,a as n,o as s}from"./chunks/framework.Caol4N0K.js";const m=JSON.parse('{"title":"Planned features","description":"","frontmatter":{},"headers":[],"relativePath":"todo.md","filePath":"todo.md","lastUpdated":null}'),o={name:"todo.md"};function r(i,a,u,d,p,c){return s(),l("div",null,a[0]||(a[0]=[e("h1",{id:"Planned-features",tabindex:"-1"},[n("Planned features "),e("a",{class:"header-anchor",href:"#Planned-features","aria-label":'Permalink to "Planned features {#Planned-features}"'},"​")],-1),e("p",null,"The following features are planned for future releases:",-1),e("ul",null,[e("li",null,[e("p",null,"Accept DEMs with missing data automatically")]),e("li",null,[e("p",null,"Classification metrics such as geomorphons")]),e("li",null,[e("p",null,"More support for multiresolution")]),e("li",null,[e("p",null,"Integration with DEM datasets/tiles")])],-1)]))}const h=t(o,[["render",r]]);export{m as __pageData,h as default};
