+++ previews/PR74/assets/chunks/@localSearchIndexroot.D1dTIRu9.js
@@ -1,0 +1,1 @@
+const e='{"documentCount":80,"nextId":80,"documentIds":{"0":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Changelog","1":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.7.3-2026-03-07","2":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Added","3":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Fixed","4":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Changed","5":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Deprecated","6":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.7.2-2026-03-03","7":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Added-2","8":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.7.1-2025-09-03","9":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Fixed-2","10":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.7.0-2025-03-21","11":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Added-3","12":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Changed-2","13":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Fixed-3","14":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.6.0-2023-08-25","15":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Changed-3","16":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Added-4","17":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.5.2-2023-08-14","18":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Changed-4","19":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.5.1-2023-01-12","20":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Added-5","21":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.5.0-2022-11-02","22":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Added-6","23":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Changed-5","24":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.4.0-2022-10-17","25":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Added-7","26":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Changed-6","27":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.3.2-2022-09-07","28":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Added-8","29":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Fixed-4","30":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.3.1-2022-06-15","31":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Added-9","32":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.3.0-2022-02-01","33":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Added-10","34":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Changed-7","35":"/Geomorphometry.jl/previews/PR74/CHANGELOG#0.2.0-2021-11-22","36":"/Geomorphometry.jl/previews/PR74/CHANGELOG#Added-11","37":"/Geom
[…]

[…]
