### 2026-05-18 73e9a3f
Make examples reproducable (#74)
* Make examples reproducable.

* Make examples reproducable.
