+++ docs/src/usage.md
@@ -12,9 +12,9 @@ CairoMakie.activate!(type = "png")
 A = GeoArrays.read("saba.tif") # hide
 B = GeoArrays.read("saba_dsm.tif") # hide
 mask = ismissing.(A) # hide
-# dem = coalesce(A, NaN).A # hide
-dtm = coalesce(A, NaN) # hide
-dsm = coalesce(B, NaN) # hide
+# dem = coalesce(A, NaN32).A # hide
+dtm = coalesce(A, NaN32) # hide
+dsm = coalesce(B, NaN32) # hide
 # ndem = GeoArrays.flipud!(deepcopy(dtm))
 
 fdtm = deepcopy(dtm)
@@ -61,7 +61,56 @@ end
 In Geomorphometry.jl we provide a set of tools to analyze and visualize the shape of the Earth. The package is designed to be fast, flexible, and easy to use.
 Moreover, we have implemented several algorithms for a common operation so that you can choose the one that best fits your needs.
 
-In these pages we will use the elevation model of [Saba](https://en.wikipedia.org/wiki/Saba_(island)) to showcase the different categories of operations that are available in Geomorphometry.jl. However, any DEM should work, most algorithms accept an `AbstractMatrix{<:Real}`. Note that you thus need to remove missing values from your DEM, which is ideally done by interpolation, but a quick `coalesce.(dem, NaN)` will work too (but will cause NaN gaps in your output).
+In these pages we will use the elevation model of [Saba](https://en.wikipedia.org/wiki/Saba_(island)) to showcase the different categories of operations that are available in Geomorphometry.jl. However, any DEM should work, most algorithms accept an `AbstractMatrix{<:Real}`. Note that you thus need to remove missing values from your DEM, which is ideally done by interpolation, but a quick `coalesce.(dem, NaN32)` will work too (but will cause NaN gaps in your output). Below are the steps to reproduce the example plots.
+
+::: tabs
+
+== Downloading
+```julia
+using Downloads
+
+dtm_fn = abspath("saba.tif")
+isfile(dtm_fn) || Downloads.download(
+    "https://github.com/Deltares/Geomorphometry.jl/releases/download/v0.6.0/saba.tif",
+    dtm_fn,
+)
+dsm_fn = abspath("saba_dsm.tif")
+isfile(dsm_fn) || Downloads.download(
+    "https://github.com/Deltares/Geomorphometry.jl/releases/download/v0.6.0/saba_dsm.tif",
+    dsm_fn,
+)
+```
+
+== Reading (using Rasters.jl)
+```julia
+using Rasters
+import ArchGDAL
+
+dtm = Raster(dtm_fn)
+dsm = Raster(dsm_fn)
+mask = ismissing.(dtm)  # used for hiding nodata areas in plots
+dtm = replace_missing(dtm, NaN32)
+dsm = replace_missing(dsm, NaN32)
+```
+
+== Reading (using GeoArrays.jl)
+```julia
+using GeoArrays
+
+dtm = GeoArrays.read(dtm_fn)
+dsm = GeoArrays.read(dsm_fn)
+mask = ismissing.(dtm)  # used for hiding nodata areas in plots
+dtm = coalesce(dtm, NaN32)
+dsm = coalesce(dsm, NaN32)
+```
+
+== Plotting
+```julia
+using CairoMakie  # or GLMakie, Plots
+
+heatmap(dtm)
+```
+:::
 
 ::: details
 
@@ -249,8 +298,8 @@ Related to the relative position is the category related to the view (to a horiz
 
[…]
