### 2026-05-23 4da2f6e
Add doi for three papers.
