+++ docs/src/refs.bib
@@ -270,6 +270,7 @@
   month     = mar,
   journal   = {--},
   publisher = {EarthArXiv},
+  doi       = {10.31223/x57x8x},
   url       = {https://eartharxiv.org/repository/view/12015/},
   urldate   = {2026-03-07},
   abstract  = {Current flood risk models applied at national and global scales do not---or only partially---take levees into account, resulting in inaccurate flood inundation maps. While levees are important assets in natural hazard risk assessments, accurate information in the public domain about the location and height of these embankments is often missing. Remote sensing data---such as global digital elevation models (DEMs)---contain this information, in theory. However, their low resolution and vertical biases (caused by the canopy and infrastructure) complicate the extraction of levees. We present a new DEM processing method to identify and extract levees from any high-resolution DEM as continuous features, including related attributes such as the protected area and volume. As a test case, we use CopernicusDEM GLO-30 and local DEMs. Validation against reference data from the United States and the Netherlands yields a recall of up to 88\% on a local DEM and up to 51\% when applied to the 30 m resolution global DSM. The method also identifies other water-retaining barriers such as undocumented dams. Incorporating levees derived from these methods into flood risk models should enable practitioners to construct more accurate (coastal) flood inundation maps on national and global scales.},
@@ -367,6 +368,7 @@
   title     = {Quantitative analysis of land surface topography},
   author    = {Zevenbergen, Lyle W and Thorne, Colin R},
   journal   = {Earth surface processes and landforms},
+  doi       = {10.1002/esp.3290120107},
   volume    = {12},
   number    = {1},
   pages     = {47--56},
@@ -437,6 +439,7 @@
   publisher  = {Elsevier},
   address    = {Amsterdam, Netherlands ; Oxford, UK ; Boston [Mass.]},
   isbn       = {978-0-12-374345-9},
+  doi        = {10.1016/s0166-2481(08)x0001-7},
   lccn       = {GB400.4 .G44 2009},
   keywords   = {Aufsatzsammlung,Computer simulation,Data processing,Digital soil mapping,Digital techniques,Digitales Gelandemodell,Geoinformationssystem,Geomorphologie,Geomorphology,Image processing,Mathematical models,Methodology,Morphometrie,Soil science},
   annotation = {OCLC: ocn271770579}
