### 2026-05-26 5e24d86
fix horizon angle check to block e.g. 24
