### 2026-05-29 e746970
fix horizon model
