### 2026-06-01 81bfbbd
Update shorthand doi.
