+++ docs/src/refs.bib
@@ -218,7 +218,7 @@
   number   = {1},
   pages    = {13--29},
   issn     = {0022-1694},
-  doi      = {10/bs5wdq},
+  doi      = {10.1016/j.jhydrol.2011.03.051},
   urldate  = {2021-07-21},
   abstract = {This paper introduces a new terrain model named HAND, and reports on the calibration and validation of landscape classes representing soil environments in Amazonia, which were derived using it. The HAND model normalizes topography according to the local relative heights found along the drainage network, and in this way, presents the topology of the relative soil gravitational potentials, or local draining potentials. The HAND model has been demonstrated to show a high correlation with the depth of the water table, providing an accurate spatial representation of soil water environments. Normalized draining potentials can be classified according to the relative vertical flowpath-distances to the nearest drainages, defining classes of soil water environments. These classes have been shown to be comparable and have verifiable and reproducible hydrological significance across the studied catchment and for surrounding ungauged catchments. The robust validation of this model over an area of 18,000km2 in the lower Rio Negro catchment has demonstrated its capacity to map expansive environments using only remotely acquired topography data as inputs. The classified HAND model has also preliminarily demonstrated robustness when applied to ungauged catchments elsewhere with contrasting geologies, geomorphologies and soil types. The HAND model and the derived soil water maps can help to advance physically based hydrological models and be applied to a host of disciplines that focus on soil moisture and ground water dynamics. As an original assessment of soil water in the landscape, the HAND model explores the synergy between digital topography data and terrain modeling, presenting an opportunity for solving many difficult problems in hydrology.},
   langid   = {english},
@@ -403,7 +403,7 @@
   number     = {1},
   pages      = {65--98},
   issn       = {0036-1445, 1095-7200},
-  doi        = {10/f9wkpj},
+  doi        = {10.1137/141000671},
   url        = {https://epubs.siam.org/doi/10.1137/141000671},
   urldate    = {2020-12-21},
   langid     = {english},
