### 2026-06-05 414df0b
CompatHelper: bump compat for Rasters in [extras] to 0.15, (keep existing compat)
