+++ Project.toml
@@ -41,7 +41,7 @@ LocalFilters = "1.2"
 OffsetArrays = "1.10"
 PaddedViews = "0.5"
 QuickHeaps = "0.2"
-Rasters = "0.13, 0.14"
+Rasters = "0.13, 0.14, 0.15"
 StaticArrays = "1"
 Statistics = "1"
 StatsBase = "0.33, 0.34"
