+++ previews/PR79/assets/index.md.DlWsxqrJ.js
@@ -1,0 +1,1 @@
+import{_ as t,c as a,a5 as i,o as l}from"./chunks/framework.NdXva-i1.js";const m=JSON.parse('{"title":"","description":"","frontmatter":{"layout":"home","hero":{"name":"Geomorphometry.jl","tagline":"Analyzing and visualizing the shape of the Earth","image":{"src":"logo.svg","alt":"Geomorphometry"},"actions":[{"theme":"brand","text":"Get Started","link":"/installation"},{"theme":"alt","text":"View on Github","link":"https://github.com/Deltares/Geomorphometry.jl"},{"theme":"alt","text":"API Reference","link":"/reference"}]},"features":[{"title":"Common operations","details":"Defines common methods for analyzing, filtering and visualizing (global) elevation models. All methods are implemented in Julia and are fast and scalable.","link":"/usage"},{"title":"Multiple algorithms","details":"Choose from multiple algorithms, ranging from different derivations of slope to multiresolution stencils. This enables the user to select the most appropriate method for their use case.","link":"/concepts"},{"title":"Seamless integration","details":"Geomorphometry.jl is fully compatible with the AbstractArray and GeoInterface.jl ecosystems. This enables plotting, operations and analysis using the full power of the Julia ecosystem."}]},"headers":[],"relativePath":"index.md","filePath":"index.md","lastUpdated":null}'),n={name:"index.md"};function o(s,e,r,p,c,h){return l(),a("div",null,e[0]||(e[0]=[i('<h2 id="Functionality" tabindex="-1">Functionality <a class="header-anchor" href="#Functionality" aria-label="Permalink to &quot;Functionality {#Functionality}&quot;">​</a></h2><ul><li><p>Terrain filters, such as Progressive Morphological Filters (PMF, SMF) and Skewness balancing</p></li><li><p>Geospatial cost (friction) operations that mimic PCRaster. These functions should however be more Julian, extensible and scale better.</p></li><li><p>Visualization, such as Perceptually Shaded Slope Map (PSSM)</p></li><li><p>Terrain analysis functions, such as slope, aspect, roughness, Topographic Position Index (TPI), Terrain Ruggedness Index (TRI), curvature and hillslope.</p></li></ul><h2 id="Installation" tabindex="-1">Installation <a class="header-anchor" href="#Installation" aria-label="Permalink to &quot;Installation {#Installation}&quot;">​</a></h2><p>The package can be installed with the Julia package manager. From the Julia REPL, type <code>]</code> to enter the Pkg REPL mode and run:</p><div class="language- vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang"></span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span>pkg&gt; add Geomorphometry</span></span></code></pre></div>',5)]))}const u=t(n,[["render",o]]);export{m as __pageData,u as default};
