+++ previews/PR79/assets/research.md.Cvx_vRyk.js
@@ -1,0 +1,1 @@
+import{_ as r,c as a,a5 as o,o as t}from"./chunks/framework.NdXva-i1.js";const m=JSON.parse('{"title":"Research","description":"","frontmatter":{},"headers":[],"relativePath":"research.md","filePath":"research.md","lastUpdated":null}'),s={name:"research.md"};function l(n,e,i,c,h,d){return t(),a("div",null,e[0]||(e[0]=[o('<h1 id="Research" tabindex="-1">Research <a class="header-anchor" href="#Research" aria-label="Permalink to &quot;Research {#Research}&quot;">​</a></h1><p>We use Geomorphometry.jl for research, please cite the software if you do too.</p><ul><li>Pronk, M. (Mar 2026). <a href="https://zenodo.org/records/18851929" target="_blank" rel="noreferrer"><em>Geomorphometry.Jl</em></a>. Zenodo. Accessed on Mar 3, 2026.</li></ul><h2 id="Papers-using-Geomorphometry.jl" tabindex="-1">Papers using Geomorphometry.jl <a class="header-anchor" href="#Papers-using-Geomorphometry.jl" aria-label="Permalink to &quot;Papers using Geomorphometry.jl {#Papers-using-Geomorphometry.jl}&quot;">​</a></h2><ul><li><p>Pronk, M.; Gawehn, M.; Eleveld, M. and Ledoux, H. (2026). <a href="https://eartharxiv.org/repository/view/12015/" target="_blank" rel="noreferrer"><em>Automated Levee Detection in Digital Elevation Models</em></a>. –. Accessed on Mar 7, 2026. Preprint, not peer-reviewed.</p></li><li><p>Pronk, M.; Hooijer, A.; Eilander, D.; Haag, A.; de Jong, T.; Vousdoukas, M.; Vernimmen, R.; Ledoux, H. and Eleveld, M. (2024). <a href="https://www.nature.com/articles/s41597-024-03091-9" target="_blank" rel="noreferrer"><em>DeltaDTM: A Global Coastal Digital Terrain Model</em></a>. <a href="https://doi.org/10.1038/s41597-024-03091-9" target="_blank" rel="noreferrer">Scientific Data <strong>11</strong>, 273</a>. Accessed on Mar 7, 2024.</p></li></ul>',5)]))}const u=r(s,[["render",l]]);export{m as __pageData,u as default};

[…]
