### 2026-06-10 4621b04
Fresh up with Aqua (#79)
* Add Manifest-v*.toml to .gitignore

Co-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>

* Freshen up with Aqua.

---------

Co-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
