+++ src/pmf.jl
@@ -1,7 +1,9 @@
 """
-    B, flags = progressive_morphological_filter(A; ωₘ, slope, dhₘ, dh₀, cellsize, adjust, erode)
+    B, flags = progressive_morphological_filter(A; ωₘ, slope, dhₘ, dh₀, cellsize, circular, adjust, erode)
 
 Applies the progressive morphological filter by [Zhang (2003)](@cite keqizhangProgressiveMorphologicalFilter2003) to `A`.
+
+A 3D `A` with a singleton third dimension is also accepted and filtered as its 2D slice.
 
 # Output
 - `B::Array{T,2}` Maximum allowable values
@@ -16,6 +18,9 @@ Afterwards, one can retrieve the resulting mask for `A` by `A .<= B` or `flags .
 - `dhₘ::Real=2.5` Maximum elevation threshold [m]
 - `dh₀::Real=0.2` Initial elevation threshold [m]
 - `cellsize::Real=1.` Cellsize in [m]
+- `circular::Bool=false` Use a circular (disk) structuring element instead of a square one
+- `adjust::Bool=false` Adjust elevation thresholds for the initial window
+- `erode::Bool=false` Apply an erosion step before the opening
 """
 function progressive_morphological_filter(
     A::AbstractMatrix{<:Real};
@@ -129,9 +134,9 @@ function _pmf(
     B, flags
 end
 
-function pmf(A::AbstractArray{<:Real, 3}; kwargs...)
+function progressive_morphological_filter(A::AbstractArray{<:Real, 3}; kwargs...)
     size(A, 3) == 1 || throw(ArgumentError("Only singleton 3rd dimension allowed"))
-    pmf(view(A, :, :, 1); kwargs...)
+    progressive_morphological_filter(view(A, :, :, 1); kwargs...)
 end
 
 function pmf2(
