+++ src/spread.jl
@@ -6,23 +6,29 @@ const Δ = CartesianIndex(1, 1)
 "Spread algorithms."
 abstract type SpreadMethod end
 
+"""
+    Tomlin()
+
+Friction-distance [`spread`](@ref) method based on a priority queue (Dijkstra-like search)
+as described by [Tomlin (1983)](@cite tomlin1983digital). This is the default method.
+"""
 Base.@kwdef struct Tomlin <: SpreadMethod end
 
 """
-    spread(::Tomlin, points::Matrix{<:Real}, initial::Matrix{<:Real}, friction::Matrix{<:Real}; res=1, limit=Inf, method=Tomlin())
+    spread(::Tomlin, locations::Vector{CartesianIndex{2}}, initial::AbstractMatrix{<:Real}, friction::AbstractMatrix{<:Real}; cellsize=cellsize(friction), limit=typemax(T))
 
-Total friction distance spread from `points` as described by [Tomlin (1983)](@cite tomlin1983digital).
+Total friction distance spread from `locations` as described by [Tomlin (1983)](@cite tomlin1983digital).
 This is also the method implemented by [PCRaster](https://pcraster.geo.uu.nl/pcraster/4.0.2/doc/manual/op_spread.html).
 
 # Output
-- `Array{Float64,2}` Total friction distance
+- `Matrix{<:Real}` Total friction distance
 
 # Arguments
-- `points::Matrix{<:Real}` Input Array
-- `initial::Matrix{<:Real}` Initial values of the result
-- `friction::Matrix{<:Real}` Resolution of cell size
-- `res=1` Resolution or cell size
-- `limit=Inf` Initial fill value
+- `locations::Vector{CartesianIndex{2}}` Source cells to spread from
+- `initial::AbstractMatrix{<:Real}` Initial values of the result at the source cells
+- `friction::AbstractMatrix{<:Real}` Friction (cost) per cell
+- `cellsize=cellsize(friction)` Cell size, used to scale distances
+- `limit=typemax(T)` Initial fill value for unreached cells
 """
 function spread(
     ::Tomlin,
@@ -78,26 +84,33 @@ function spread!(pq, mask, result, friction, zone, distances)
     end
 end
 
+"""
+    Eastman(; iterations=3)
+
+Friction-distance [`spread`](@ref) method using the pushbroom approach of
+[Eastman (1989)](@cite eastman1989pushbroom). Scales better (linearly) than [`Tomlin`](@ref),
+but may require more `iterations` for maze-like, uncrossable obstacles.
+"""
 Base.@kwdef struct Eastman <: SpreadMethod
     iterations::Int = 3
 end
 
 """
-    spread(::Eastman, points::Matrix{<:Real}, initial::Matrix{<:Real}, friction::Matrix{<:Real}; res=1, limit=Inf, iterations=3)
+    spread(::Eastman, locations::Vector{CartesianIndex{2}}, initial::AbstractMatrix{<:Real}, friction::AbstractMatrix{<:Real}; cellsize=cellsize(friction), limit=Inf)
 
 Pushbroom method for friction costs as discussed by [Eastman (1989)](@cite eastman1989pushbroom).
 This method should scale better (linearly) than the [Tomlin (1983)](@cite tomlin1983digital) method, but can require more
 `iterations` than set by default (3) in the case of maze-like, uncrossable obstacles.
 
 # Output
-- `Array{Float64,2}` Total friction distance
+- `Matrix{<:Real}` Total friction distance
 
 # Arguments
-- `points::Matrix{<:Real}` Input Array
-- `initial::Matrix{<:Real}` Factor to exaggerate elevation
-- `friction::Matrix{<:Real}` Resolution of cell size
-- `res=1` Resolution or cell size
[…]
