+++ Project.toml
@@ -30,13 +30,16 @@ GeomorphometryGeoArraysExt = "GeoArrays"
 GeomorphometryRastersExt = ["Rasters", "ArchGDAL"]
 
 [compat]
+Aqua = "0.8"
 ArchGDAL = "0.10"
 DataStructures = "0.18, 0.19"
 Distances = "0.10"
 Eikonal = "0.1.1"
+ExplicitImports = "1.15"
 FillArrays = "0.12, 0.13, 1"
 GeoArrays = "0.9"
 KernelAbstractions = "0.9.40"
+LinearAlgebra = "1"
 LocalFilters = "1.2"
 OffsetArrays = "1.10"
 PaddedViews = "0.5"
@@ -46,14 +49,17 @@ StaticArrays = "1"
 Statistics = "1"
 StatsBase = "0.33, 0.34"
 Stencils = "0.3.4"
+Test = "1"
 julia = "1.10"
 
 [extras]
+Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 ArchGDAL = "c9ce4bd3-c3d5-55b8-8973-c0e20141b8c3"
 Eikonal = "a6aab1ba-8f88-4217-b671-4d0788596809"
+ExplicitImports = "7d51a73a-1435-4ff3-83d9-f097790105c7"
 GeoArrays = "2fb1d81b-e6a0-5fc5-82e6-8e06903437ab"
 Rasters = "a3a2b9e3-a471-40c9-b274-f788e487c689"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "GeoArrays", "Rasters", "ArchGDAL", "Eikonal"]
+test = ["Test", "GeoArrays", "Rasters", "ArchGDAL", "Eikonal", "Aqua", "ExplicitImports"]

[…]
