### 2026-06-11 7938130
add missing_elevation keyword and default to 0.0
