+++ src/horizon.jl
@@ -1,5 +1,5 @@
 """
-    horizon_angle(dem; method=GridSweep(), cellsize=cellsize(dem))
+    horizon_angle(dem; method=GridSweep(), cellsize=cellsize(dem), missing_elevation=0.0)
 
 Compute horizon angles for each cell in a DEM.
 
@@ -13,6 +13,9 @@ the horizontal (sheltered), negative angles indicate terrain falling below (expo
 # Keywords
 - `directions`: 16 by default, can be 4, 8, 16, 32, ...
 - `cellsize`: Cell size as `(row_size, col_size)` tuple
+- `missing_elevation`: Elevation substituted for NaN cells along a sweep ray
+  (default `0.0`). Use `Inf` to make NaN fully occluding. Cells whose own
+  elevation is NaN always produce `NaN` in the output.
 
 # Returns
 3D array of size `(rows, cols, directions)` with horizon angles in degrees.
@@ -20,57 +23,54 @@ the horizontal (sheltered), negative angles indicate terrain falling below (expo
 function horizon_angle(dem::AbstractMatrix{<:Real};
     directions::Int = 16,
     cellsize = cellsize(dem),
+    missing_elevation::Real = 0.0,
 )
     ispow2(directions) && directions >= 4 ||
         throw(ArgumentError("directions must be 4, 8, 16, 32, ..., got $directions"))
+    me = Float64(missing_elevation)
     if directions == 4
-        _horizon_angle_cardinal(dem, cellsize)
+        _horizon_angle_cardinal(dem, cellsize, me)
     elseif directions == 8
-        _horizon_angle_8(dem, cellsize)
+        _horizon_angle_8(dem, cellsize, me)
     else
-        _horizon_angle_rotated(dem, directions, cellsize)
+        _horizon_angle_rotated(dem, directions, cellsize, me)
     end
 end
 
 # KernelAbstractions kernels for line sweeps
 
 # Sweep from row edge (top or bottom) - index gives column
-@kernel function _sweep_row_edge_kernel!(out, @Const(dem), start_row::Int, step_r::Int, step_c::Int, dist::Float64, nrows::Int, ncols::Int)
+@kernel function _sweep_row_edge_kernel!(out, @Const(dem), start_row::Int, step_r::Int, step_c::Int, dist::Float64, nrows::Int, ncols::Int, missing_elevation::Float64)
     col = @index(Global)
-    _sweep_line_ka!(out, dem, start_row, col, step_r, step_c, dist, nrows, ncols)
+    _sweep_line_ka!(out, dem, start_row, col, step_r, step_c, dist, nrows, ncols, missing_elevation)
 end
 
 # Sweep from col edge (left or right) - index gives row
-@kernel function _sweep_col_edge_kernel!(out, @Const(dem), row_offset::Int, start_col::Int, step_r::Int, step_c::Int, dist::Float64, nrows::Int, ncols::Int)
+@kernel function _sweep_col_edge_kernel!(out, @Const(dem), row_offset::Int, start_col::Int, step_r::Int, step_c::Int, dist::Float64, nrows::Int, ncols::Int, missing_elevation::Float64)
     idx = @index(Global)
-    _sweep_line_ka!(out, dem, idx + row_offset, start_col, step_r, step_c, dist, nrows, ncols)
+    _sweep_line_ka!(out, dem, idx + row_offset, start_col, step_r, step_c, dist, nrows, ncols, missing_elevation)
 end
 
[…]
