### 2026-06-11 bf8ec89
use similar for KernelAbstractions code so Raster/GeoArray are returned
