### 2026-06-11 e6da10e
Fix horizon angle calculation for distant terrain  (#75)
* fix horizon angle check to block e.g. 24

* fix horizon model

* cleanup comment

* add missing_elevation keyword and default to 0.0
