+++ test/horizon.jl
@@ -40,6 +40,9 @@ using Test
     @testset "invalid directions" begin
         flat = ones(5, 5)
         @test_throws ArgumentError horizon_angle(flat; directions=6)
+        # Multiples of 8 that aren't powers of 2 (e.g. 24, 40) must also reject.
+        @test_throws ArgumentError horizon_angle(flat; directions=24)
+        @test_throws ArgumentError horizon_angle(flat; directions=40)
     end
 
     @testset "sloping terrain" begin
@@ -55,6 +58,64 @@ using Test
         # E-W should be flat (0°)
         @test h[5, 5, 2] ≈ 0.0 atol=1e-5  # E
         @test h[5, 5, 4] ≈ 0.0 atol=1e-5  # W
+    end
+
+    @testset "distant peak beyond flat foreground" begin
+        # 10 m peak at the north edge, 9 cells of flat ground in front.
+        dem = fill(100.0, 10, 5)
+        dem[1, :] .= 110.0
+        h = horizon_angle(dem; directions = 4, cellsize = (1.0, 1.0))
+        # N, E, S, W. From (10, 3) the peak is 9 cells north.
+        @test h[10, 3, 1] ≈ atand(10 / 9) atol = 1e-3
+        @test h[10, 3, 2] ≈ 0.0 atol = 1e-5
+        @test h[10, 3, 4] ≈ 0.0 atol = 1e-5
+    end
+
+    @testset "angle to a spike decays with distance" begin
+        # 10 m spike at row 2; observers south of it must see it shrink.
+        dem = fill(100.0, 10, 5)
+        dem[2, :] .= 110.0
+        h = horizon_angle(dem; directions = 4, cellsize = (1.0, 1.0))
+        @test h[5, 3, 1] ≈ atand(10 / 3) atol = 1e-3
+        @test h[10, 3, 1] ≈ atand(10 / 8) atol = 1e-3
+        @test h[10, 3, 1] < h[5, 3, 1]
+    end
+
+    @testset "missing cells use missing_elevation (default 0)" begin
+        # 200 m peak at row 1, NaN row at row 2, flat 100 m elsewhere.
+        dem = fill(100.0, 10, 5)
+        dem[1, :] .= 200.0
+        dem[2, :] .= NaN
+        h = horizon_angle(dem; directions = 4, cellsize = (1.0, 1.0))
+        # NaN row treated as 0 m (below observer) so peak 9 cells north dominates.
+        @test h[10, 3, 1] ≈ atand(100 / 9) atol = 1e-3
+        @test isnan(h[2, 3, 1])
+    end
+
+    @testset "missing_elevation = Inf fully occludes" begin
+        dem = fill(100.0, 10, 5)
+        dem[1, :] .= 200.0
+        dem[2, :] .= NaN
+        h = horizon_angle(dem; directions = 4, cellsize = (1.0, 1.0), missing_elevation = Inf)
+        @test h[10, 3, 1] ≈ 90.0 atol = 1e-3
+    end
+
+    @testset "missing_elevation custom value" begin
+        # NaN row 8 cells north of observer; substitute 150 m → 50 m rise.
+        dem = fill(100.0, 10, 5)
+        dem[2, :] .= NaN
+        h = horizon_angle(dem; directions = 4, cellsize = (1.0, 1.0), missing_elevation = 150.0)
+        @test h[10, 3, 1] ≈ atand(50 / 8) atol = 1e-3
+    end
+
+    @testset "horizon scales with cellsize" begin
+        # 5 m bump at the south edge, observer 4 cells north.
+        dem = fill(100.0, 5, 3)
+        dem[5, :] .= 105.0
[…]
