### 2026-06-11 d788cfd
use `similar` for KernelAbstractions code so Raster/GeoArray are returned (#80)
* use similar for KernelAbstractions code so Raster/GeoArray are returned

* move where horizon tests run for extensions
