### 2026-06-11 62420f9
Add viewshed and total_viewshed methods.
