+++ src/horizon.jl
@@ -1,5 +1,5 @@
 """
-    horizon_angle(dem; directions=16, cellsize=cellsize(dem), missing_elevation=0.0)
+    horizon_angle(dem; directions=16, cellsize=cellsize(dem), observer_height=0.0, missing_elevation=0.0)
 
 Compute horizon angles for each cell in a DEM.
 
@@ -13,6 +13,8 @@ the horizontal (sheltered), negative angles indicate terrain falling below (expo
 # Keywords
 - `directions`: 16 by default, can be 4, 8, 16, 32, ...
 - `cellsize`: Cell size as `(row_size, col_size)` tuple
+- `observer_height`: Height of the observer's eye above the terrain (default `0.0`). A
+  positive value raises the viewpoint, lowering the computed horizon angles.
 - `missing_elevation`: Elevation substituted for NaN cells along a sweep ray
   (default `0.0`). Use `Inf` to make NaN fully occluding. Cells whose own
   elevation is NaN always produce `NaN` in the output.
@@ -23,17 +25,19 @@ the horizontal (sheltered), negative angles indicate terrain falling below (expo
 function horizon_angle(dem::AbstractMatrix{<:Real};
     directions::Int = 16,
     cellsize = cellsize(dem),
+    observer_height::Real = 0.0,
     missing_elevation::Real = 0.0,
 )
     ispow2(directions) && directions >= 4 ||
         throw(ArgumentError("directions must be 4, 8, 16, 32, ..., got $directions"))
     me = Float64(missing_elevation)
+    op = HorizonAngle(Float64(observer_height))
     if directions == 4
-        _horizon_angle_cardinal(dem, cellsize, me)
+        _sweep_directions_cardinal(op, dem, cellsize, me)
     elseif directions == 8
-        _horizon_angle_8(dem, cellsize, me)
+        _sweep_directions_8(op, dem, cellsize, me)
     else
-        _horizon_angle_rotated(dem, directions, cellsize, me)
+        _sweep_directions_rotated(op, dem, directions, cellsize, me)
     end
 end
 
@@ -44,43 +48,89 @@ end
 _alloc_directions(dem, T, ndirs) =
     similar(dem, T, size(dem, 1), size(dem, 2), ndirs)
 
+# Sweep operations. Singleton structs selected at the call site so the inner
+# branch specializes and the unused path is eliminated (no runtime if/else).
+abstract type SweepOp end
+
+# Per-cell backward look-back producing the horizon angle (a reduction). `height`
+# raises the observer's eye above the terrain (metres).
+struct HorizonAngle <: SweepOp
+    height::Float64
+end
+
+# Per-observer running-max along the line, accumulating the fraction of cells
+# each observer can see (used by `total_viewshed`). `height` raises the observer's
+# eye above the terrain (metres), so cells on a uniform/grazing slope are seen at a
+# progressively shallower depression angle and count as visible.
+struct TotalVisibility <: SweepOp
+    height::Float64
+end
 # KernelAbstractions kernels for line sweeps
 
 # Sweep from row edge (top or bottom) - index gives column
-@kernel function _sweep_row_edge_kernel!(out, @Const(dem), start_row::Int, step_r::Int, step_c::Int, dist::Float64, nrows::Int, ncols::Int, missing_elevation::Float64)
[…]
