### 2026-06-16 98b6b47
Add viewshed and total_viewshed methods. (#81)
