+++ docs/src/refs.bib
@@ -58,7 +58,7 @@
   title       = {Principles of {{Geographical Information Systems}}},
   author      = {Burrough, P. A. and McDonnell, Rachael A. and Lloyd, Christopher D.},
   year        = {2015},
-  publisher   = {OUP Oxford},
+  publisher   = {Oxford University Press},
   abstract    = {Geographical data are used in so many aspects of our lives today, ranging from disaster relief operations through to finding directions on our mobile phones. We can all be data collectors, adding locational information as we capture digitally our day-to-day experiences. Geographical Information Systems (GIS) are the software tools that facilitate this, turning the raw data into useful information that can help us understand our worlds better. Principles of Geographical Information Systems presents a thorough overview of the subject, exploring both the theoretical basis of GIS, and their use in practice. It explains how data on the world are converted into digital form and the analytical capabilities used to bring understanding to a range of areas of interest and issues. Spatial data are usually based on two, dichotomous paradigms: exactly defined entities in space, such as land parcels and urban structures, or the continuous variation of single attributes, such as temperature or rainfall. The adoption of one or the other influences how the geographical data are structured in the GIS and the types of analysis possible. A further area of focus in the book concerns the problems of data quality and how statistical errors in spatial data can affect the results of spatial modelling based on the two paradigms of space. Fuzzy logic and continuous classification methods are presented as methods for linking the two spatial paradigms. The book concludes with an overview of current developments and trends in providing spatial data to an ever-expanding global community of users. Online Resource Centre The Online Resource Centre to accompany Principles of Geographical Information Systems features: For students: {$\cdot$} Links to useful software and software providers for GIS {$\cdot$} Datasets with instructions of how to repeat analyses presented in the book using widely-used software For registered adopters of the book: {$\cdot$} Figures from the book, available for download},
   googlebooks = {kvoJCAAAQBAJ},
   isbn        = {978-0-19-874284-5},
@@ -86,9 +86,9 @@
 
 @inproceedings{eastman1989pushbroom,
   title     = {Pushbroom algorithms for calculating distances in raster grids},
-  author    = {Eastman, J Ronald},
   booktitle = {Proceedings, Autocarto},
   volume    = {9},
+  author    = {Eastman, J. Ronald},
   pages     = {288--297},
   year      = {1989}
 }
@@ -130,9 +130,12 @@
 
 @article{jensonExtractingTopographicStructure1988,
   title    = {Extracting {{Topographic Structure}} from {{Digital Elevation Data}} for {{Geographic Information System Analysis}}},
-  author   = {Jenson, S K and Domingue, J O},
+  author   = {Jenson, S. K. and Domingue, J. O.},
   year     = {1988},
-  journal  = {PHOTOGRAMMETRIC ENGINEERING},
+  journal  = {Photogrammetric Engineering and Remote Sensing},
+  volume   = {54},
+  number   = {11},
+  pages    = {1593--1600},
   abstract = {Software tools have been developed at the U.S. Geological Survey's EROS Data Center to extract topographic structure and to delineate watersheds and overland flow paths from digital elevation models. The tools are special purpose FORTRAN programs interfaced with general-purpose raster and vector spatial analysis and relational data base management packages.},
   langid   = {english}
 }
@@ -187,11 +190,14 @@
   langid    = {english},
[…]
