### 2026-06-24 2694947
Add JOSS badge
