### 2026-06-24 d8507e9
Update CFF with JOSS doi
