### 2026-06-24 696f697
build based on 2694947
