+++ dev/assets/chunks/theme.Di1yBLTR.js
@@ -1,0 +1,2 @@
+const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/chunks/VPLocalSearchBox.DaVV5noE.js","assets/chunks/framework.De9caxpi.js"])))=>i.map(i=>d[i]);
+import{d as b,o as a,c as u,r as p,n as A,a as X,t as N,b as g,w as v,e as _,T as Pe,_ as $,u as Je,i as _t,f as bt,g as Se,h as V,j as f,k as i,l as Y,m as _e,p as C,q as z,s as de,v as G,x as Ve,y as Le,z as gt,A as kt,B as Z,F as I,C as O,D as Xe,E as pe,G as k,H as R,I as Ye,J as fe,K as q,L as ee,M as $t,N as Ze,O as be,P as we,Q as Qe,R as ve,S as yt,U as Pt,V as St,W as xe,X as Te,Y as Vt,Z as Lt,$ as wt,a0 as Tt,a1 as et,a2 as Ct,a3 as Mt,a4 as Fe}from"./framework.De9caxpi.js";const Nt=b({__name:"VPBadge",props:{text:{},type:{default:"tip"}},setup(n){return(e,t)=>(a(),u("span",{class:A(["VPBadge",e.type])},[p(e.$slots,"default",{},()=>[X(N(e.text),1)])],2))}}),It={key:0,class:"VPBackdrop"},At=b({__name:"VPBackdrop",props:{show:{type:Boolean}},setup(n){return(e,t)=>(a(),g(Pe,{name:"fade"},{default:v(()=>[e.show?(a(),u("div",It)):_("",!0)]),_:1}))}}),Et=$(At,[["__scopeId","data-v-b06cdb19"]]),L=Je;function Bt(n,e){let t,s=!1;return()=>{t&&clearTimeout(t),s?t=setTimeout(n,e):(n(),(s=!0)&&setTimeout(()=>s=!1,e))}}function ge(n){return/^\//.test(n)?n:`/${n}`}function Ce(n){const{pathname:e,search:t,hash:s,protocol:o}=new URL(n,"http://a.com");if(_t(n)||n.startsWith("#")||!o.startsWith("http")||!bt(e))return n;const{site:r}=L(),l=e.endsWith("/")||e.endsWith(".html")?n:n.replace(/(?:(^\.+)\/)?.*$/,`$1${e.replace(/(\.md)?$/,r.value.cleanUrls?"":".html")}${t}${s}`);return Se(l)}function ne({correspondingLink:n=!1}={}){const{site:e,localeIndex:t,page:s,theme:o,hash:r}=L(),l=V(()=>{var c,m;return{label:(c=e.value.locales[t.value])==null?void 0:c.label,link:((m=e.value.locales[t.value])==null?void 0:m.link)||(t.value==="root"?"/":`/${t.value}/`)}});return{localeLinks:V(()=>Object.entries(e.value.locales).flatMap(([c,m])=>l.value.label===m.label?[]:{text:m.label,link:Ht(m.link||(c==="root"?"/":`/${c}/`),o.value.i18nRouting!==!1&&n,s.value.relativePath.slice(l.value.link.length-1),!e.value.cleanUrls)+r.value})),currentLang:l}}function Ht(n,e,t,s){return e?n.replace(/\/$/,"")+ge(t.replace(/(^|\/)index\.md$/,"$1").replace(/\.md$/,s?".html":"")):n}const Ot={class:"NotFound"},Dt={class:"code"},Ft={class:"title"},Ut={class:"quote"},Rt={class:"action"},jt=["href","aria-label"],zt=b({__name:"NotFound",setup(n){
[…]
