+++ dev/assets/validation.md.origvMtj.lean.js
@@ -1,0 +1,7 @@
+import{_ as n,c as p,a5 as d,G as a,w as t,j as i,a as e,B as k,o as r}from"./chunks/framework.De9caxpi.js";const o="/Geomorphometry.jl/dev/assets/krdtexc.DqjcziFy.png",g="/Geomorphometry.jl/dev/assets/mcongdf.CUq-rert.png",E="/Geomorphometry.jl/dev/assets/hzzdbcm.DSheVUgv.png",y="/Geomorphometry.jl/dev/assets/bjkwtph.j4kvbTJH.png",u="/Geomorphometry.jl/dev/assets/dkoptmh.Cm8Rf6xm.png",m="/Geomorphometry.jl/dev/assets/owqtvbe.D-LRB5-7.png",c="/Geomorphometry.jl/dev/assets/mwxjlpu.Cvl-pgRW.png",b="/Geomorphometry.jl/dev/assets/cmmbetc.DrE1q7OI.png",C="/Geomorphometry.jl/dev/assets/sxtaozd.CMMUjAq7.png",F="/Geomorphometry.jl/dev/assets/pihwbqk.B6ILuj_3.png",v="/Geomorphometry.jl/dev/assets/bjiznrq.Di-r3_Np.png",x="/Geomorphometry.jl/dev/assets/rsxdzyi.BF7NKjh7.png",f="/Geomorphometry.jl/dev/assets/jmikkkn.CxKsrEpd.png",j="/Geomorphometry.jl/dev/assets/jmikkkn.CxKsrEpd.png",B="/Geomorphometry.jl/dev/assets/bjiznrq.Di-r3_Np.png",w="/Geomorphometry.jl/dev/assets/rsxdzyi.BF7NKjh7.png",A="/Geomorphometry.jl/dev/assets/kuvktlb.DED2al9h.png",D="/Geomorphometry.jl/dev/assets/xvgicjg.1DIjzyhj.png",G="/Geomorphometry.jl/dev/assets/ecvlzwx.CI-YTQQs.png",_="/Geomorphometry.jl/dev/assets/ecvlzwx.CI-YTQQs.png",T="/Geomorphometry.jl/dev/assets/cctiqht.BhbWh8sC.png",L="/Geomorphometry.jl/dev/assets/cctiqht.BhbWh8sC.png",H="/Geomorphometry.jl/dev/assets/ecvlzwx.CI-YTQQs.png",Z="/Geomorphometry.jl/dev/assets/ecvlzwx.CI-YTQQs.png",N=JSON.parse('{"title":"Validation","description":"","frontmatter":{},"headers":[],"relativePath":"validation.md","filePath":"validation.md","lastUpdated":null}'),z={name:"validation.md"};function P(q,s,O,V,S,I){const l=k("PluginTabsTab"),h=k("PluginTabs");return r(),p("div",null,[s[24]||(s[24]=d(`<h1 id="Validation" tabindex="-1">Validation <a class="header-anchor" href="#Validation" aria-label="Permalink to &quot;Validation {#Validation}&quot;">​</a></h1><p>This chapter compares our output with that of <a href="https://gdal.org/en/stable/programs/gdaldem.html" target="_blank" rel="noreferrer">gdaldem</a>.</p><p>We test with the same elevation model as that in the <a href="./usage">Usage</a> page, and also include an upside down version of it, to ensure that the algorithms are robust to different axes conventions.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#005CC5;--shiki-d
[…]

[…]
