### 2026-06-29 c9bd07d
CompatHelper: bump compat for Rasters in [weakdeps] to 0.15, (keep existing compat) (#77)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
