+++ dev/assets/concepts.md.BZbIrsmr.js
@@ -1,4 +1,4 @@
-import{_ as p,c as d,a5 as k,G as l,w as t,j as i,a as s,B as n,o as r}from"./chunks/framework.De9caxpi.js";const g="/Geomorphometry.jl/dev/assets/vualheh.Dw5qi7Tw.png",o="/Geomorphometry.jl/dev/assets/djexvuj.Cm8Rf6xm.png",E="/Geomorphometry.jl/dev/assets/jvdgqgv.CiddrJvp.png",y="/Geomorphometry.jl/dev/assets/jacvvjq.BbjN0-LY.png",c="/Geomorphometry.jl/dev/assets/zuyjcfx.CIKj522l.png",u="/Geomorphometry.jl/dev/assets/oskqfwg.PNNbvaLU.png",C="/Geomorphometry.jl/dev/assets/wtyiuid.BgVM-P3m.png",F="/Geomorphometry.jl/dev/assets/gtbwsnx.Ds5R6Z50.png",m="/Geomorphometry.jl/dev/assets/rixiavo.BMd0ixZB.png",b="/Geomorphometry.jl/dev/assets/mgeqrfj.ZJ1g3DCB.png",v="/Geomorphometry.jl/dev/assets/xmvagdb.BTIjrNCM.png",B="/Geomorphometry.jl/dev/assets/ikxvwpg.ChSy4ryr.png",D="/Geomorphometry.jl/dev/assets/irarrme.B8A4wtsI.png",f="/Geomorphometry.jl/dev/assets/tfbvjvy.DpneZ2_w.png",S=JSON.parse('{"title":"Concepts","description":"","frontmatter":{},"headers":[],"relativePath":"concepts.md","filePath":"concepts.md","lastUpdated":null}'),x={name:"concepts.md"};function j(A,a,w,_,G,M){const h=n("PluginTabsTab"),e=n("PluginTabs");return r(),d("div",null,[a[15]||(a[15]=k('<h1 id="Concepts" tabindex="-1">Concepts <a class="header-anchor" href="#Concepts" aria-label="Permalink to &quot;Concepts {#Concepts}&quot;">​</a></h1><p>In Geomorphometry.jl we provide a set of tools to analyse and visualize the shape of the Earth. The package is designed to be fast, flexible, and easy to use. It is built with the following concepts in mind:</p><h2 id="Geospatially-aware" tabindex="-1">Geospatially aware <a class="header-anchor" href="#Geospatially-aware" aria-label="Permalink to &quot;Geospatially aware {#Geospatially-aware}&quot;">​</a></h2><p>With package extensions on <a href="https://github.com/evetion/GeoArrays.jl" target="_blank" rel="noreferrer">GeoArrays.jl</a> and <a href="https://github.com/rafaqz/Rasters.jl" target="_blank" rel="noreferrer">Rasters.jl</a> geospatial data is automatically handled to set the correct cellsize, even for geographical DEMs.</p>',4)),l(e,null,{default:t(()=>[l(h,{label:"Rasters (projected)"},{default:t(()=>a[0]||(a[0]=[i("div",{class:"language-julia vp-adaptive-theme"},[i("button",{title:"Copy Code",class:"copy"}),i("span",{class:"lang"},"julia"),i("pre",{class:"shiki shiki-themes github-light github-dark vp-code",tabindex:"0"},[i("code",null,[i("span",{class:"line"},[i("span",{style:{"--shiki-light":"#24292E","--shiki-dark":"#E1E4E8"}},"r "),i("span",{style:{"--shiki-light":"#D73A49","--shiki-dark":"#F97583"}},"="),i("span",{style:{"--shiki-light":"#005CC5","-
[…]
