+++ previews/PR70/assets/experimental.md.BnUAGnm-.lean.js
@@ -1,1 +1,0 @@
-import{_ as r,c as l,j as s,a,G as i,a5 as o,w as n,B as p,o as d}from"./chunks/framework.CT0Wd1gt.js";const x=JSON.parse('{"title":"Experimental","description":"","frontmatter":{},"headers":[],"relativePath":"experimental.md","filePath":"experimental.md","lastUpdated":null}'),c={name:"experimental.md"},m={class:"jldocstring custom-block",open:""};function h(k,e,u,b,g,y){const t=p("Badge");return d(),l("div",null,[e[4]||(e[4]=s("h1",{id:"Experimental",tabindex:"-1"},[a("Experimental "),s("a",{class:"header-anchor",href:"#Experimental","aria-label":'Permalink to "Experimental {#Experimental}"'},"​")],-1)),e[5]||(e[5]=s("div",{class:"warning custom-block github-alert"},[s("p",{class:"custom-block-title"},"Methods here are experimental, not yet stable and may change or even be removed in future releases."),s("p")],-1)),s("details",m,[s("summary",null,[e[0]||(e[0]=s("a",{id:"Geomorphometry.skbr",href:"#Geomorphometry.skbr"},[s("span",{class:"jlbinding"},"Geomorphometry.skbr")],-1)),e[1]||(e[1]=a()),i(t,{type:"info",class:"jlObjectType jlFunction",text:"Function"})]),e[3]||(e[3]=o('<div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">mask </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> skbr</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(A; iterations</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">10</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">)</span></span></code></pre></div><p>Applies recursive skewness balancing by <a href="/Geomorphometry.jl/previews/PR70/bibliography#bartelsThresholdfreeObjectGround2010">Bartels et al. (2010)</a> to <code>A</code>. Applies <code>skewness_balancing</code> <code>iterations</code> times to the object (non-terrain) mask, as to include more (sloped) terrain.</p><p><strong>Output</strong></p><ul><li><code>mask::BitMatrix</code> Mask of allowed values</li></ul>',4)),i(t,{type:"info",class:"source-link",text:"source"},{default:n(()=>e[2]||(e[2]=[s("a",{href:"https://github.com/Deltares/Geomorphometry.jl/blob/7ef4e5d7de7844a0ea67a98874b6bfccd5978047/src/skew.jl#LL85-L94",target:"_blank",rel:"noreferrer"},"source",-1)])),_:1})])])}const E=r(c,[["render",h]]);export{x as __pageData,E as default};

[…]
