+++ Project.toml
@@ -1,6 +1,6 @@
 name = "Geomorphometry"
 uuid = "714e3e49-7933-471a-9334-4a6a65a92f36"
-version = "0.7.3"
+version = "0.7.4"
 authors = ["Maarten Pronk <git@evetion.nl>", "Deltares"]
 
 [deps]
