+++ dev/assets/validation.md.CJLHME7g.js
@@ -1,7 +1,0 @@
-import{_ as k,c as p,a5 as d,G as a,w as t,j as i,a as e,B as n,o as r}from"./chunks/framework.De9caxpi.js";const o="/Geomorphometry.jl/dev/assets/iijgljl.DqjcziFy.png",g="/Geomorphometry.jl/dev/assets/maudqnw.CUq-rert.png",E="/Geomorphometry.jl/dev/assets/xkzyebr.DSheVUgv.png",y="/Geomorphometry.jl/dev/assets/ekckmrg.j4kvbTJH.png",u="/Geomorphometry.jl/dev/assets/jwfjhhu.Cm8Rf6xm.png",m="/Geomorphometry.jl/dev/assets/rhyubgq.D-LRB5-7.png",c="/Geomorphometry.jl/dev/assets/cpkblhh.Cvl-pgRW.png",C="/Geomorphometry.jl/dev/assets/hgkyiqt.DrE1q7OI.png",b="/Geomorphometry.jl/dev/assets/cpsprou.CMMUjAq7.png",F="/Geomorphometry.jl/dev/assets/ajirnac.B6ILuj_3.png",v="/Geomorphometry.jl/dev/assets/vnolpdl.Di-r3_Np.png",f="/Geomorphometry.jl/dev/assets/cdcmrnl.BF7NKjh7.png",x="/Geomorphometry.jl/dev/assets/faueqsq.CxKsrEpd.png",j="/Geomorphometry.jl/dev/assets/faueqsq.CxKsrEpd.png",B="/Geomorphometry.jl/dev/assets/vnolpdl.Di-r3_Np.png",A="/Geomorphometry.jl/dev/assets/cdcmrnl.BF7NKjh7.png",D="/Geomorphometry.jl/dev/assets/gjlbqbw.DED2al9h.png",G="/Geomorphometry.jl/dev/assets/btfrfrr.1DIjzyhj.png",w="/Geomorphometry.jl/dev/assets/gsvtudm.CI-YTQQs.png",_="/Geomorphometry.jl/dev/assets/gsvtudm.CI-YTQQs.png",T="/Geomorphometry.jl/dev/assets/nrldtcu.BhbWh8sC.png",L="/Geomorphometry.jl/dev/assets/nrldtcu.BhbWh8sC.png",H="/Geomorphometry.jl/dev/assets/gsvtudm.CI-YTQQs.png",Z="/Geomorphometry.jl/dev/assets/gsvtudm.CI-YTQQs.png",Q=JSON.parse('{"title":"Validation","description":"","frontmatter":{},"headers":[],"relativePath":"validation.md","filePath":"validation.md","lastUpdated":null}'),q={name:"validation.md"};function P(O,s,V,S,I,M){const l=n("PluginTabsTab"),h=n("PluginTabs");return r(),p("div",null,[s[24]||(s[24]=d(`<h1 id="Validation" tabindex="-1">Validation <a class="header-anchor" href="#Validation" aria-label="Permalink to &quot;Validation {#Validation}&quot;">​</a></h1><p>This chapter compares our output with that of <a href="https://gdal.org/en/stable/programs/gdaldem.html" target="_blank" rel="noreferrer">gdaldem</a>.</p><p>We test with the same elevation model as that in the <a href="./usage">Usage</a> page, and also include an upside down version of it, to ensure that the algorithms are robust to different axes conventions.</p><div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#005CC5;--shiki-dark:
[…]
