+++ src/geomorphon.jl
@@ -1,0 +1,223 @@
+"""
+    Landform
+
+A landform class of the geomorphon classification, as returned by [`geomorphon`](@ref).
+It is stored as an integer, with the same category numbers as GRASS `r.geomorphon`:
+
+- [`Flat`](@ref Landform) (1): no relief in any direction
+- [`Peak`](@ref Landform) (2): higher than its surroundings in all directions
+- [`Ridge`](@ref Landform) (3): a convex, elongated crest
+- [`Shoulder`](@ref Landform) (4): convex break in slope, from flat to falling
+- [`Spur`](@ref Landform) (5): convex (nose) slope
+- [`Slope`](@ref Landform) (6): a planar, tilted surface
+- [`Hollow`](@ref Landform) (7): concave (colluvial) slope
+- [`Footslope`](@ref Landform) (8): concave break in slope, from falling to flat
+- [`Valley`](@ref Landform) (9): a concave, elongated trough
+- [`Pit`](@ref Landform) (10): lower than its surroundings in all directions
+
+[`Undefined`](@ref Landform) (0) marks cells that could not be classified, because their
+own elevation is missing or no cell along their lines of sight has a valid elevation.
+"""
+struct Landform <: Integer
+    value::UInt8
+    # Inner constructor, so that no default converting constructor is generated that
+    # would be ambiguous with the `Integer` constructors in Base.
+    Landform(value::Integer) = new(value)
+end
+
+const Undefined = Landform(0)
+const Flat = Landform(1)
+const Peak = Landform(2)
+const Ridge = Landform(3)
+const Shoulder = Landform(4)
+const Spur = Landform(5)
+const Slope = Landform(6)
+const Hollow = Landform(7)
+const Footslope = Landform(8)
+const Valley = Landform(9)
+const Pit = Landform(10)
+
+const _LANDFORM_NAMES = (
+    "Undefined",
+    "Flat",
+    "Peak",
+    "Ridge",
+    "Shoulder",
+    "Spur",
+    "Slope",
+    "Hollow",
+    "Footslope",
+    "Valley",
+    "Pit",
+)
+
+Base.show(io::IO, l::Landform) = print(io, _LANDFORM_NAMES[l.value + 1])
+Base.:(==)(a::Landform, b::Landform) = a.value == b.value
+Base.:(<)(a::Landform, b::Landform) = a.value < b.value
+Base.Int(l::Landform) = Int(l.value)
+Base.UInt8(l::Landform) = l.value
+# The category numbers run from `Undefined` to `Pit`; wrapper types use these to pick a
+# missing value for a classification.
+Base.typemin(::Type{Landform}) = Undefined
+Base.typemax(::Type{Landform}) = Pit
+# Compare and hash as the category number it stands for.
+Base.promote_rule(::Type{Landform}, ::Type{T}) where {T <: Integer} = T
+Base.hash(l::Landform, h::UInt) = hash(l.value, h)
+
+# Landform for each (number of negative, number of positive) ternary pattern element
+# count, indexed as `_LANDFORMS[nminus + 1, nplus + 1]`. Combinations that sum to more
+# than the eight directions are impossible and marked `Undefined`. This is the lookup
+# table of [jasiewiczGeomorphonsPatternRecognition2013](@cite).
+const _LANDFORMS = @SMatrix [
+#   +0        +1        +2        +3        +4        +5        +6        +7        +8
+    Flat      Flat      Flat      Footslope Footslope Valley    Valley    Valley    Pit        # -0
+    Flat      Flat      Footslope Footslope Footslope Valley    Valley    Valley    Undefined  # -1
+    Flat      Shoulder  Slope     Slope     Hollow    Hollow    Valley    Undefined Undefined  # -2
[…]
