+++ docs/src/refs.bib
@@ -128,6 +128,18 @@
   langid     = {english}
 }
 
+@article{jasiewiczGeomorphonsPatternRecognition2013,
+  title    = {Geomorphons \textemdash{} a Pattern Recognition Approach to Classification and Mapping of Landforms},
+  author   = {Jasiewicz, Jaros{\l}aw and Stepinski, Tomasz F.},
+  year     = {2013},
+  journal  = {Geomorphology},
+  volume   = {182},
+  pages    = {147--156},
+  issn     = {0169-555X},
+  doi      = {10.1016/j.geomorph.2012.11.005},
+  langid   = {english}
+}
+
 @article{jensonExtractingTopographicStructure1988,
   title    = {Extracting {{Topographic Structure}} from {{Digital Elevation Data}} for {{Geographic Information System Analysis}}},
   author   = {Jenson, S K and Domingue, J O},

[…]
