+++ docs/src/refs.bib
@@ -129,7 +129,7 @@
 }
 
 @article{jasiewiczGeomorphonsPatternRecognition2013,
-  title    = {Geomorphons \textemdash{} a Pattern Recognition Approach to Classification and Mapping of Landforms},
+  title    = {Geomorphons---a Pattern Recognition Approach to Classification and Mapping of Landforms},
   author   = {Jasiewicz, Jaros{\l}aw and Stepinski, Tomasz F.},
   year     = {2013},
   journal  = {Geomorphology},
