### 2026-09-17 036f62d
Fix plotting.
