+++ docs/Project.toml
@@ -16,4 +16,4 @@ Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 Geomorphometry = {path = ".."}
 
 [compat]
-Makie = "0.22"
+Makie = "0.24"
