+++ previews/PR87/assets/concepts.md.C2SDFfKk.js
@@ -1,0 +1,39 @@
+import{_ as p,c as r,a5 as k,G as l,w as t,j as i,a as s,B as n,o as d}from"./chunks/framework.DGabKWNo.js";const g="/Geomorphometry.jl/previews/PR87/assets/xwozfrz.Dw5qi7Tw.png",o="/Geomorphometry.jl/previews/PR87/assets/amlebqb.Cm8Rf6xm.png",E="/Geomorphometry.jl/previews/PR87/assets/hdsnvzn.CiddrJvp.png",y="/Geomorphometry.jl/previews/PR87/assets/qeecgck.BbjN0-LY.png",c="/Geomorphometry.jl/previews/PR87/assets/cvvqtqk.CIKj522l.png",u="/Geomorphometry.jl/previews/PR87/assets/hoagfgl.PNNbvaLU.png",C="/Geomorphometry.jl/previews/PR87/assets/hbbzfoe.BgVM-P3m.png",F="/Geomorphometry.jl/previews/PR87/assets/jkgpwku.Ds5R6Z50.png",m="/Geomorphometry.jl/previews/PR87/assets/hdtmwrl.BMd0ixZB.png",b="/Geomorphometry.jl/previews/PR87/assets/qgfamzs.ZJ1g3DCB.png",v="/Geomorphometry.jl/previews/PR87/assets/nzsztju.BTIjrNCM.png",B="/Geomorphometry.jl/previews/PR87/assets/fjcvhzb.ChSy4ryr.png",D="/Geomorphometry.jl/previews/PR87/assets/uvbscty.B8A4wtsI.png",f="/Geomorphometry.jl/previews/PR87/assets/pphloox.DpneZ2_w.png",P=JSON.parse('{"title":"Concepts","description":"","frontmatter":{},"headers":[],"relativePath":"concepts.md","filePath":"concepts.md","lastUpdated":null}'),x={name:"concepts.md"};function A(j,a,w,_,G,M){const h=n("PluginTabsTab"),e=n("PluginTabs");return d(),r("div",null,[a[15]||(a[15]=k('<h1 id="Concepts" tabindex="-1">Concepts <a class="header-anchor" href="#Concepts" aria-label="Permalink to &quot;Concepts {#Concepts}&quot;">​</a></h1><p>In Geomorphometry.jl we provide a set of tools to analyse and visualize the shape of the Earth. The package is designed to be fast, flexible, and easy to use. It is built with the following concepts in mind:</p><h2 id="Geospatially-aware" tabindex="-1">Geospatially aware <a class="header-anchor" href="#Geospatially-aware" aria-label="Permalink to &quot;Geospatially aware {#Geospatially-aware}&quot;">​</a></h2><p>With package extensions on <a href="https://github.com/evetion/GeoArrays.jl" target="_blank" rel="noreferrer">GeoArrays.jl</a> and <a href="https://github.com/rafaqz/Rasters.jl" target="_blank" rel="noreferrer">Rasters.jl</a> geospatial data is automatically handled to set the correct cellsize, even for geographical DEMs.</p>',4)),l(e,null,{default:t(()=>[l(h,{label:"Rasters (projected)"},{default:t(()=>a[0]||(a[0]=[i("div",{class:"language-julia vp-adaptive-theme"},[i("button",{title:"Copy Code",class:"copy"}),i("span",{class:"lang"},"julia"),i("pre",{class:"shiki shiki-themes github-light github-dark vp-code",tabindex:"0"},[i("code",null,[i("span",{class:"line"},[i("span",{style:{"--shiki-light":"#24292E","--shiki-dark":"#E1E4E8"}},"r "),i("span"
[…]
