### 2026-09-17 660102f
Add geomorphon landform classification (#87)
* Add geomorphon landform classification

Reduce the eight lines of sight around each cell to a ternary pattern of
rising, flat and falling directions, and map the number of rising and
falling directions onto one of ten Landform classes, following Jasiewicz
and Stepinski (2013). Landform subtypes Integer, like FlowDirection, so
that the result rides along in a Raster or GeoArray.

Fixes #39.

Co-Authored-By: Claude Opus 5 (1M context) <noreply@anthropic.com>

* Fix bib.

* Fix plotting.

* Fix plots.

---------

Co-authored-by: Claude Opus 5 (1M con
[…]
