+++ src/Geomorphometry.jl
@@ -40,6 +40,7 @@ include("skew.jl")
 include("flowdir.jl")
 include("hydrology.jl")
 include("horizon.jl")
+include("geomorphon.jl")
 
 export ZevenbergenThorne, Horn, MDG
 export D8, DInf, FD8
@@ -65,5 +66,8 @@ export filldepressions,
     height_above_nearest_drainage
 export FlowDirection, LDD, D8D
 export horizon_angle, sky_view_factor, viewshed, total_viewshed
+export geomorphon
+export Landform,
+    Undefined, Flat, Peak, Ridge, Shoulder, Spur, Slope, Hollow, Footslope, Valley, Pit
 
 end # module
