+++ dev/assets/chunks/@localSearchIndexroot.ZOu2w687.js
@@ -1,1 +1,0 @@
-const e='{"documentCount":85,"nextId":85,"documentIds":{"0":"/Geomorphometry.jl/dev/CHANGELOG#Changelog","1":"/Geomorphometry.jl/dev/CHANGELOG#Unreleased","2":"/Geomorphometry.jl/dev/CHANGELOG#0.7.4-2026-07-01","3":"/Geomorphometry.jl/dev/CHANGELOG#Added","4":"/Geomorphometry.jl/dev/CHANGELOG#Fixed","5":"/Geomorphometry.jl/dev/CHANGELOG#Changed","6":"/Geomorphometry.jl/dev/CHANGELOG#0.7.3-2026-03-07","7":"/Geomorphometry.jl/dev/CHANGELOG#Added-2","8":"/Geomorphometry.jl/dev/CHANGELOG#Fixed-2","9":"/Geomorphometry.jl/dev/CHANGELOG#Changed-2","10":"/Geomorphometry.jl/dev/CHANGELOG#Deprecated","11":"/Geomorphometry.jl/dev/CHANGELOG#0.7.2-2026-03-03","12":"/Geomorphometry.jl/dev/CHANGELOG#Added-3","13":"/Geomorphometry.jl/dev/CHANGELOG#0.7.1-2025-09-03","14":"/Geomorphometry.jl/dev/CHANGELOG#Fixed-3","15":"/Geomorphometry.jl/dev/CHANGELOG#0.7.0-2025-03-21","16":"/Geomorphometry.jl/dev/CHANGELOG#Added-4","17":"/Geomorphometry.jl/dev/CHANGELOG#Changed-3","18":"/Geomorphometry.jl/dev/CHANGELOG#Fixed-4","19":"/Geomorphometry.jl/dev/CHANGELOG#0.6.0-2023-08-25","20":"/Geomorphometry.jl/dev/CHANGELOG#Changed-4","21":"/Geomorphometry.jl/dev/CHANGELOG#Added-5","22":"/Geomorphometry.jl/dev/CHANGELOG#0.5.2-2023-08-14","23":"/Geomorphometry.jl/dev/CHANGELOG#Changed-5","24":"/Geomorphometry.jl/dev/CHANGELOG#0.5.1-2023-01-12","25":"/Geomorphometry.jl/dev/CHANGELOG#Added-6","26":"/Geomorphometry.jl/dev/CHANGELOG#0.5.0-2022-11-02","27":"/Geomorphometry.jl/dev/CHANGELOG#Added-7","28":"/Geomorphometry.jl/dev/CHANGELOG#Changed-6","29":"/Geomorphometry.jl/dev/CHANGELOG#0.4.0-2022-10-17","30":"/Geomorphometry.jl/dev/CHANGELOG#Added-8","31":"/Geomorphometry.jl/dev/CHANGELOG#Changed-7","32":"/Geomorphometry.jl/dev/CHANGELOG#0.3.2-2022-09-07","33":"/Geomorphometry.jl/dev/CHANGELOG#Added-9","34":"/Geomorphometry.jl/dev/CHANGELOG#Fixed-5","35":"/Geomorphometry.jl/dev/CHANGELOG#0.3.1-2022-06-15","36":"/Geomorphometry.jl/dev/CHANGELOG#Added-10","37":"/Geomorphometry.jl/dev/CHANGELOG#0.3.0-2022-02-01","38":"/Geomorphometry.jl/dev/CHANGELOG#Added-11","39":"/Geomorphometry.jl/dev/CHANGELOG#Changed-8","40":"/Geomorphometry.jl/dev/CHANGELOG#0.2.0-2021-11-22","41":"/Geomorphometry.jl/dev/CHANG
[…]
