+++ docs/src/refs.bib
@@ -437,7 +437,7 @@
 
 @article{zaksekSkyViewFactorRelief2011,
   title    = {Sky-View Factor as a Relief Visualization Technique},
-  author   = {Zak{\v s}ek, Klemen and O{\v s}tir, Kri{\v s}tof and Kokalj, {\v Z}iga},
+  author   = {Zakšek, Klemen and Oštir, Krištof and Kokalj, Žiga},
   year     = {2011},
   journal  = {Remote Sensing},
   volume   = {3},
