+++ previews/PR88/assets/guide.md.DjAc9Oa7.js
@@ -1,0 +1,1 @@
+import{_ as t,c as s,j as a,a as d,o as r}from"./chunks/framework.CNLi__IA.js";const f=JSON.parse('{"title":"Guides","description":"","frontmatter":{},"headers":[],"relativePath":"guide.md","filePath":"guide.md","lastUpdated":null}'),i={name:"guide.md"};function o(n,e,l,u,c,p){return r(),s("div",null,e[0]||(e[0]=[a("h1",{id:"Guides",tabindex:"-1"},[d("Guides "),a("a",{class:"header-anchor",href:"#Guides","aria-label":'Permalink to "Guides {#Guides}"'},"​")],-1)]))}const _=t(i,[["render",o]]);export{f as __pageData,_ as default};
