+++ previews/PR88/assets/CHANGELOG.md.CUbQybBh.lean.js
@@ -1,0 +1,1 @@
+import{_ as d,c as a,a5 as o,o as i}from"./chunks/framework.CNLi__IA.js";const f=JSON.parse('{"title":"Changelog","description":"","frontmatter":{},"headers":[],"relativePath":"CHANGELOG.md","filePath":"CHANGELOG.md","lastUpdated":null}'),l={name:"CHANGELOG.md"};function r(t,e,n,h,c,s){return i(),a("div",null,e[0]||(e[0]=[o('<h1 id="Changelog" tabindex="-1">Changelog <a class="header-anchor" href="#Changelog" aria-label="Permalink to &quot;Changelog {#Changelog}&quot;">​</a></h1><p>All notable changes to this project will be documented in this file.</p><h2 id="Unreleased" tabindex="-1">[Unreleased] <a class="header-anchor" href="#Unreleased" aria-label="Permalink to &quot;[Unreleased] {#Unreleased}&quot;">​</a></h2><h2 id="Added" tabindex="-1">Added <a class="header-anchor" href="#Added" aria-label="Permalink to &quot;Added {#Added}&quot;">​</a></h2><ul><li>Added <code>geomorphon</code> to classify each cell of a DEM into one of ten <code>Landform</code> classes.</li></ul><h2 id="0.7.4-2026-07-01" tabindex="-1">[0.7.4] - 2026-07-01 <a class="header-anchor" href="#0.7.4-2026-07-01" aria-label="Permalink to &quot;[0.7.4] - 2026-07-01 {#0.7.4-2026-07-01}&quot;">​</a></h2><h2 id="Added-2" tabindex="-1">Added <a class="header-anchor" href="#Added-2" aria-label="Permalink to &quot;Added {#Added-2}&quot;">​</a></h2><ul><li>Added <code>viewshed</code> (single-observer line-of-sight visibility) and <code>total_viewshed</code> (normalized visibility index).</li></ul><h2 id="Fixed" tabindex="-1">Fixed <a class="header-anchor" href="#Fixed" aria-label="Permalink to &quot;Fixed {#Fixed}&quot;">​</a></h2><ul><li><code>sky_view_factor</code> and <code>horizon_angle</code> now work on for longer distance.</li></ul><h2 id="Changed" tabindex="-1">Changed <a class="header-anchor" href="#Changed" aria-label="Permalink to &quot;Changed {#Changed}&quot;">​</a></h2><ul><li><code>sky_view_factor</code> and <code>horizon_angle</code> now use similar(input) for output, e.g. keeping a Raster or GeoArray.</li></ul><h2 id="0.7.3-2026-03-07" tabindex="-1">[0.7.3] - 2026-03-07 <a class="header-anchor" href="#0.7.3-2026-03-07" aria-label="Permalink to &quot;[0.7.3] - 2026-03-07 {#0.7.3-2026-03-07}&quot;">​</a></h2><h2 id="Added-3" tabindex="-1">Added <a class="header-anchor" href="#Added-3" aria-label="Permalink to &quot;Added {#Added-3}&quot;">​</a></h2><ul><li><p>Added methods <code>depression_depth</code>, <code>depression_volume</code>, <code>drainage_potential</code> and <code>percentile_elevation</code>.</p></li><li><p>Added <code>FlowDirection</code> that wraps integer encodings for flow direction.</p></li></ul><h2 id="Fixed-2" tabindex="-1">Fixed <a class="header-anchor" href="#Fixed-2" aria-label="Permalink to &quot;Fixed {#Fixed-2}&quot;">​</a></h2><ul><li>No more double flow accumulation on edges (cells wi
[…]
