+++ previews/PR88/assets/CONTRIBUTING.md.MNGf9shE.lean.js
@@ -1,0 +1,1 @@
+import{_ as t,c as i,a5 as a,o as s}from"./chunks/framework.CNLi__IA.js";const c=JSON.parse('{"title":"Contributing","description":"","frontmatter":{},"headers":[],"relativePath":"CONTRIBUTING.md","filePath":"CONTRIBUTING.md","lastUpdated":null}'),o={name:"CONTRIBUTING.md"};function r(n,e,l,p,u,h){return s(),i("div",null,e[0]||(e[0]=[a('<h1 id="Contributing" tabindex="-1">Contributing <a class="header-anchor" href="#Contributing" aria-label="Permalink to &quot;Contributing {#Contributing}&quot;">​</a></h1><p>Thank you for your interest in contributing to Geomorphometry.jl! This document provides guidelines and information for contributors.</p><h2 id="Reporting-Issues" tabindex="-1">Reporting Issues <a class="header-anchor" href="#Reporting-Issues" aria-label="Permalink to &quot;Reporting Issues {#Reporting-Issues}&quot;">​</a></h2><p>If you encounter a bug or have a problem with the software, please <a href="https://github.com/Deltares/Geomorphometry.jl/issues/new" target="_blank" rel="noreferrer">open an issue</a> on GitHub. When reporting issues, please include:</p><ul><li><p>A clear and descriptive title.</p></li><li><p>A minimal reproducible example (MRE) that demonstrates the problem (please share data that you&#39;re using as input)</p></li><li><p>The version of Geomorphometry.jl you are using (<code>] status Geomorphometry</code> in the Julia REPL).</p></li><li><p>Your Julia version (<code>versioninfo()</code>).</p></li><li><p>Any relevant error messages or stack traces.</p></li></ul><h2 id="Feature-Requests" tabindex="-1">Feature Requests <a class="header-anchor" href="#Feature-Requests" aria-label="Permalink to &quot;Feature Requests {#Feature-Requests}&quot;">​</a></h2><p>We welcome suggestions for new features or improvements. Please <a href="https://github.com/Deltares/Geomorphometry.jl/issues/new" target="_blank" rel="noreferrer">open an issue</a> and describe:</p><ul><li><p>The feature you would like to see.</p></li><li><p>The use case or problem it would solve.</p></li><li><p>Any relevant references (papers, existing implementations, etc.).</p></li></ul><h2 id="Contributing-Code" tabindex="-1">Contributing Code <a class="header-anchor" href="#Contributing-Code" aria-label="Permalink to &quot;Contributing Code {#Contributing-Code}&quot;">​</a></h2><h3 id="Getting-Started" tabindex="-1">Getting Started <a class="header-anchor" href="#Getting-Started" aria-label="Permalink to &quot;Getting Started {#Getting-Started}&quot;">​</a></h3><ol><li><p>Fork the repository on GitHub.</p></li><li><p>Clone your fork locally:</p><div class="language-bash vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">bash</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#6F42C1;--shiki-dark:#B392F0;">git</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> clone</span><span style="--shiki-light:#032F62;--shiki-dark:#9ECBFF;"> https://github.com/your-username/Geomorphometry.jl.git</span></span></code></pre></div></li><li><p>Create a new branch for your changes:</p><div class="language-bash vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">bash</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><c
[…]
