### 2026-09-19 5c170d5
Update citations and fix bugs in pmf and hydrology (#88)
* Correctness

* Fix ref.
