+++ docs/src/refs.bib
@@ -67,6 +67,19 @@
 }
 
 
+@article{cavalliEffectivenessAirborneLidar2008,
+  title    = {The Effectiveness of Airborne {{LiDAR}} Data in the Recognition of Channel-Bed Morphology},
+  author   = {Cavalli, Marco and Tarolli, Paolo and Marchi, Lorenzo and {Dalla Fontana}, Giancarlo},
+  year     = {2008},
+  journal  = {CATENA},
+  volume   = {73},
+  number   = {3},
+  pages    = {249--260},
+  issn     = {0341-8162},
+  doi      = {10.1016/j.catena.2007.11.001},
+  langid   = {english}
+}
+
 @article{costa-cabralDigitalElevationModel1994,
   title      = {Digital {{Elevation Model Networks}} ({{DEMON}}): {{A}} Model of Flow over Hillslopes for Computation of Contributing and Dispersal Areas},
   shorttitle = {Digital {{Elevation Model Networks}} ({{DEMON}})},
@@ -140,11 +153,26 @@
   langid   = {english}
 }
 
+@article{jennessCalculatingLandscapeSurface2004,
+  title    = {Calculating Landscape Surface Area from Digital Elevation Models},
+  author   = {Jenness, Jeff S.},
+  year     = {2004},
+  journal  = {Wildlife Society Bulletin},
+  volume   = {32},
+  number   = {3},
+  pages    = {829--839},
+  issn     = {0091-7648},
+  doi      = {10.2193/0091-7648(2004)032[0829:CLSAFD]2.0.CO;2},
+  langid   = {english}
+}
+
 @article{jensonExtractingTopographicStructure1988,
   title    = {Extracting {{Topographic Structure}} from {{Digital Elevation Data}} for {{Geographic Information System Analysis}}},
   author   = {Jenson, S K and Domingue, J O},
   year     = {1988},
-  journal  = {PHOTOGRAMMETRIC ENGINEERING},
+  journal  = {Photogrammetric Engineering and Remote Sensing},
+  volume   = {54},
+  number   = {11},
   abstract = {Software tools have been developed at the U.S. Geological Survey's EROS Data Center to extract topographic structure and to delineate watersheds and overland flow paths from digital elevation models. The tools are special purpose FORTRAN programs interfaced with general-purpose raster and vector spatial analysis and relational data base management packages.},
   langid   = {english}
 }
@@ -165,6 +193,17 @@
   annotation = {545 citations (Crossref) [2022-01-19]\\
                 ZSCC: NoCitationData[s0]}
 }
+@inproceedings{lloberaCalculatingInherentVisual2004,
+  title     = {Calculating the Inherent Visual Structure of a Landscape (`{{Total Viewshed}}') Using High-Throughput Computing},
+  author    = {Llobera, Marcos and Wheatley, David and Steele, James and Cox, Simon and Parchment, Oliver},
+  year      = {2004},
+  booktitle = {Beyond the Artefact: {{Digital}} Interpretation of the Past. {{Proceedings}} of {{CAA2004}}},
+  editor    = {Niccolucci, Franco and Hermon, Sorin},
+  pages     = {146--151},
+  publisher = {Archaeolingua},
+  address   = {Budapest}
+}
+
 @article{lundbladBenthicTerrainClassification2006,
   title     = {A {{Benthic Terrain Classification Scheme}} for {{American Samoa}}},
   author    = {Lundblad, Emily R. and Wright, Dawn J. and Miller, Joyce and Larkin, Emily M. and Rinehart, Ronald and Naar, David F. and Donahue, Brian T. and Anderson, S. Miles and Battista, Tim},
@@ -199,11 +238,15 @@
[…]
