### 2026-09-19 b976c7a
Fix the docs build: GeoArray flow accumulation and the renamed outlets
`flowaccumulation!` and `height_above_nearest_drainage` allocate their
direction field with `similar(dem, CartesianIndex{2})`. A `GeoArray` only
holds `Union{Missing,Number}`, so that call threw a `MethodError` and every
`@example` block in the docs that accumulates flow failed. Route the
allocation through a new `_alloc_offsets` hook, which the GeoArrays
extension answers with a plain array of the same shape; matrices, Rasters
and DGGS grids keep the `similar` default.

The Priority-Flood animation in `usage.md` still call
[…]
