+++ previews/PR83/assets/experimental.md.kvIMZtH5.lean.js
@@ -1,0 +1,1 @@
+import{_ as n,c as o,j as t,a,G as i,a5 as r,w as l,B as p,o as d}from"./chunks/framework.BfJGmQzy.js";const y=JSON.parse('{"title":"Experimental","description":"","frontmatter":{},"headers":[],"relativePath":"experimental.md","filePath":"experimental.md","lastUpdated":null}'),c={name:"experimental.md"},m={class:"jldocstring custom-block",open:""};function k(h,e,u,b,g,f){const s=p("Badge");return d(),o("div",null,[e[4]||(e[4]=t("h1",{id:"Experimental",tabindex:"-1"},[a("Experimental "),t("a",{class:"header-anchor",href:"#Experimental","aria-label":'Permalink to "Experimental {#Experimental}"'},"​")],-1)),e[5]||(e[5]=t("div",{class:"warning custom-block github-alert"},[t("p",{class:"custom-block-title"},"Methods here are experimental, not yet stable and may change or even be removed in future releases."),t("p")],-1)),t("details",m,[t("summary",null,[e[0]||(e[0]=t("a",{id:"Geomorphometry.skbr",href:"#Geomorphometry.skbr"},[t("span",{class:"jlbinding"},"Geomorphometry.skbr")],-1)),e[1]||(e[1]=a()),i(s,{type:"info",class:"jlObjectType jlFunction",text:"Function"})]),e[3]||(e[3]=r('<div class="language-julia vp-adaptive-theme"><button title="Copy Code" class="copy"></button><span class="lang">julia</span><pre class="shiki shiki-themes github-light github-dark vp-code" tabindex="0"><code><span class="line"><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">mask </span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;"> skbr</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">(A; iterations</span><span style="--shiki-light:#D73A49;--shiki-dark:#F97583;">=</span><span style="--shiki-light:#005CC5;--shiki-dark:#79B8FF;">10</span><span style="--shiki-light:#24292E;--shiki-dark:#E1E4E8;">)</span></span></code></pre></div><p>Applies recursive skewness balancing by DocumenterCitations.CitationSiteNode(&quot;bartelsThresholdfreeObjectGround2010-cite-1&quot;)</p><p>to <code>A</code>. Applies <code>skewness_balancing</code> <code>iterations</code> times to the object (non-terrain) mask, as to include more (sloped) terrain.</p><p><strong>Output</strong></p><ul><li><code>mask::BitMatrix</code> Mask of allowed values</li></ul>',5)),i(s,{type:"info",class:"source-link",text:"source"},{default:l(()=>e[2]||(e[2]=[t("a",{href:"https://github.com/Deltares/Geomorphometry.jl/blob/a6b4d7ffa2cec6d9489b1983a630328e03a6e68f/src/skew.jl#LL89-L98",target:"_blank",rel:"noreferrer"},"source",-1)])),_:1})])])}const E=n(c,[["render",k]]);export{y as __pageData,E as default};
